vera/pyproject.toml
[project]
name = "vera"
version = "1.0.2"
description = "AI Feature Test Engine"
readme = "README.md"
authors = [
    { name = "Tal Shapir", email = "talshapir@google.com" },
]
requires-python = ">=3.14"
dependencies = [
    "anyio>=4.12.0",
    "google-genai>=1.55.0",
    "platformdirs>=4.5.1",
    "pluggy>=1.6.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "tenacity>=9.1.2",
    "typer>=0.20.1",
]

keywords = [
    "vera",
    "google",
    "ai",
    "testing",
    "regression testing",
    "plugin",
    "test",
]

classifiers = [
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3 :: Only",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.10",
    "ty>=0.0.5",
]

[project.scripts]
vera = "vera.main:main"

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
pythonpath = "src"
norecursedirs = ["src/vera/create/plugin_name"]
addopts = [
    "--import-mode=importlib",
]
testpaths = [
    "tests/",
]

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true
